+++ src/services/juno/custom/custom.ts
@@ -1,64 +1,10 @@
-import { AdminResources, ProposalDetails } from "services/types";
-import { idParamToNum } from "helpers";
-import { contracts } from "constants/contracts";
+import { ProposalDetails } from "services/types";
 import { junoApi } from "..";
 import { queryContract } from "../queryContract";
-import { accountTags, adminTags } from "../tags";
-import { apCWs, getMeta } from "./helpers/admin-resource";
+import { adminTags } from "../tags";
 
 export const customApi = junoApi.injectEndpoints({
   endpoints: (builder) => ({
-    adminResources: builder.query<
-      AdminResources,
-      { user?: string; endowmentId?: string }
-    >({
-      providesTags: [
-        { type: "admin", id: adminTags.voter },
-        { type: "admin", id: adminTags.voters },
-        { type: "admin", id: adminTags.config },
-        { type: "account", id: accountTags.endowment },
-      ],
-      async queryFn(args) {
-        const numId = idParamToNum(args.endowmentId);
-        const AP = apCWs[numId];
-        /** special case for ap admin usage */
-        if (AP) {
-          const { cw3, cw4, type } = AP;
-          //skip endowment query, query hardcoded cw3 straight
-
-          const [meta, config] = await getMeta(numId, cw3, args.user);
-          return {
-            data: {
-              type: type as any,
-              id: numId,
-              cw3,
-              cw4,
-              propMeta: meta,
-              config,
-            },
-          };
-        }
-
-        const endowment = await queryContract(
-          "accEndowment",
-          contracts.accounts,
-          { id: numId }
-        );
-        const [meta, config] = await getMeta(numId, endowment.owner, args.user);
-
-        return {
-          data: {
-            type: "charity",
-            id: numId,
-            cw3: endowment.owner,
-            cw4: config.group_addr,
-            config,
-            propMeta: meta,
-            ...endowment,
-          },
-        };
-      },
-    }),
     proposalDetails: builder.query<
       ProposalDetails,
       { id?: string; cw3: string; voter: string }
@@ -92,4 +38,4 @@ export const customApi = junoApi.injectEndpoints({
   }),
 });
 
-export const { useAdminResourcesQuery, useProposalDetailsQuery } = customApi;
+export const { useProposalDetailsQuery } = customApi;
